fix (search) display result of search

+++ app/Http/Controllers/Admin/Users/UsersBackendController.php
@@ -10,6 +10,7 @@ use App\Http\Requests\Backend\UserRequest;
 use App\Models\User;
 use App\Repository\User\UserRepository;
 use App\Services\FlashMessagesServices;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\Support\Renderable;
 use Illuminate\Contracts\View\Factory;
@@ -25,15 +26,23 @@ class UsersBackendController extends Controller
     ) {
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function index(): Renderable
     {
+        $this->authorize('viewAny', User::class);
         $users = $this->repository->getElements();
 
         return view('admin.domain.users.index', compact('users'));
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function create(): Renderable
     {
+        $this->authorize('create', User::class);
         return view('admin.domain.users.create');
     }
 
@@ -46,13 +55,23 @@ class UsersBackendController extends Controller
         return redirect()->route('admins.users.index');
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function show(User $user): Renderable
     {
+        $this->authorize('view', $user);
+
         return view('admin.domain.users.show', compact('user'));
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function edit(User $user): Renderable
     {
+        $this->authorize('update', $user);
+
         return view('admin.domain.users.edit', compact('user'));
     }
 
@@ -65,8 +84,13 @@ class UsersBackendController extends Controller
         return redirect()->route('admins.users.index');
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function destroy(User $user): RedirectResponse
     {
+        $this->authorize('delete', $user);
+
         $this->repository->delete($user);
 
         $this->flashMessagesServices->success('success', "L'utilisateur a ete supprimer");
